% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{inflect}
\alias{inflect}
\title{Local extrema}
\usage{
inflect(x, threshold = 1)
}
\arguments{
\item{x}{Vector of values.}

\item{threshold}{Sensitivity threshold. (best results = 2-6)}
}
\value{
Extrema information.
}
\description{
Local extrema
}
