% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Std_RelRelief.R
\name{relRelief}
\alias{relRelief}
\title{Standard Relative Relief}
\usage{
relRelief(inRas, sc)
}
\arguments{
\item{inRas}{Input DEM. (SpatRas)}

\item{sc}{Scale of window.}
}
\value{
Relative relief raster (SpatRast)
}
\description{
Standard Relative Relief
}
