% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{rr.Average}
\alias{rr.Average}
\title{Average RR Surface Raster}
\usage{
rr.Average(inRas, inMn = 7, inMx = 21, stp = 2)
}
\arguments{
\item{inRas}{Input DEM. (SpatRast)}

\item{inMn}{Minimum scale. (default = 7)}

\item{inMx}{Maximum scale. (default = 21)}

\item{stp}{Scale step. (default = 2)}
}
\value{
Average relative relief surface raster. (SpatRast)
}
\description{
Average RR Surface Raster
}
