% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{rr.AverageLayer}
\alias{rr.AverageLayer}
\title{RR Average}
\usage{
rr.AverageLayer(inMn = 7, inMx = Id.Sc, x = D, inRas = M, save = T)
}
\arguments{
\item{inMn}{Minimum cale}

\item{inMx}{Maximum scale}

\item{inRas}{Input DEM}
}
\value{
Average RR surface
}
\description{
RR Average
}
