% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{rr.Mins.Analysis}
\alias{rr.Mins.Analysis}
\title{Rework RR.mn DFs for Interpretation}
\usage{
rr.Mins.Analysis(inMn, inRR)
}
\arguments{
\item{inMn}{Input RR.mn data. (List of lists)}

\item{inRR}{Input RR data. (List of lists)}
}
\value{
Simplified minima location. (List of lists)
}
\description{
Rework RR.mn DFs for Interpretation
}
