% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{rr.Mins}
\alias{rr.Mins}
\title{Plot multiple RR and moving avg. RR}
\usage{
rr.Mins(rr.L)
}
\arguments{
\item{rr.L}{Input RR data. (List of lists)}
}
\value{
List of lists of local minima.
}
\description{
Plot multiple RR and moving avg. RR
}
