% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{rr.Point}
\alias{rr.Point}
\title{Single scale RR at a cell/point}
\usage{
rr.Point(inRas, in.S = NA, ii = NA, jj = NA, thrsh = 0)
}
\arguments{
\item{inRas}{Input DEM. (matrix NOT raster)}

\item{in.S}{Scale.}

\item{ii}{i index.}

\item{jj}{j index.}

\item{thrsh}{Threshold elevation to calculate RR above.}
}
\value{
RR at cell/point for given scale.
}
\description{
Single scale RR at a cell/point
}
