% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{rr.Vec}
\alias{rr.Vec}
\title{RR of single scale along profile}
\usage{
rr.Vec(inRas, in.Sc = 7, in.ID = NA, thresh = 0, sea = NA, chk = 0)
}
\arguments{
\item{inRas}{Input DEM. (matrix NOT SpatRast)}

\item{in.Sc}{Scale.}

\item{in.ID}{ID value of profile.}

\item{thresh}{Threshold elevation to calculate RR above.}

\item{sea}{Direction/orientation of the sea/ocean. ("N", "E", "S", or "W")}

\item{chk}{Where to start calculating.}
}
\value{
RR vector at scale S for the given profile ID
}
\description{
RR of single scale along profile
}
