% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{tpi.Average}
\alias{tpi.Average}
\title{TPI Average}
\usage{
tpi.Average(inMn, inMx, inRas)
}
\arguments{
\item{inMn}{Minimum cale}

\item{inMx}{Maximum scale}

\item{inRas}{Input DEM}
}
\value{
Average TPI surface
}
\description{
TPI Average
}
