% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Std_RelRelief.R
\name{TPI}
\alias{TPI}
\title{Standard Topographic Position Index}
\usage{
TPI(inRas, sc)
}
\arguments{
\item{inRas}{Input DEM. (SpatRaster)}

\item{sc}{Scale of window.}
}
\value{
Relative relief raster (SpatRaster)
}
\description{
Standard Topographic Position Index
}
