% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{tri.Average}
\alias{tri.Average}
\title{TRI Average}
\usage{
tri.Average(inMn, inMx, inRas)
}
\arguments{
\item{inMn}{Minimum cale}

\item{inMx}{Maximum scale}

\item{inRas}{Input DEM}
}
\value{
Average TRI surface
}
\description{
TRI Average
}
