% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_RelRelief.R
\name{rr.Average}
\alias{rr.Average}
\title{Average RR Surface Raster}
\usage{
rr.Average(inMn = 7, inMx = Id.Sc, x = D, inRas = M, save = T)

rr.Average(inMn = 7, inMx = Id.Sc, x = D, inRas = M, save = T)
}
\arguments{
\item{inMn}{Minimum cale}

\item{inMx}{Maximum scale}

\item{inRas}{Input DEM}

\item{stp}{Scale step. (default = 2)}
}
\value{
Average relative relief surface raster. (SpatRast)

Average RR surface
}
\description{
Average RR Surface Raster

RR Average
}
